pyproject.toml
[tool.poetry]
name = "project-1-bank"
version = "0.1.0"
description = ""
authors = ["our ame <pastaeva98@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_return_any = true
exclude = '.venv'

[tool.isort]
# максимальная длина строки
line_length = 119
